Port GameServer/Bishop/Goddess/S3Relay build chain to x64
Adds Debug|x64 and Release|x64 configurations across the full
server-side dependency chain (Common, Engine, LuaLibDll, JpgLib,
KMp3LibClass, Core, Heaven, Rainbow, GameServer, Bishop, Goddess,
S3Relay, FilterText) so the game server cluster can build and run
as native 64-bit binaries while the client and payment system stay
32-bit.

Key fixes beyond straightforward build-system config:
- Ported/guarded x86 inline asm in Common, Engine and JpgLib behind
  _M_IX86, with portable C fallbacks for x64.
- Fixed wire-protocol structs in KProtocol.h (size_t -> DWORD) so
  x64 and 32-bit processes exchange identical byte layouts.
- Fixed pointer-truncation bugs surfaced by real runtime testing:
  _findfirst() search handles stored in `long` instead of
  intptr_t (KSortScript.cpp, DBDumpLoad.cpp), and
  iCoreServerShell::GetGameData/OperationRequest passing pointers
  through `unsigned int` params instead of UINT_PTR/INT_PTR
  (CoreServerShell.h/.cpp, ~70 call sites in KSOServer.cpp) - both
  caused access-violation crashes on x64 that took crash-dump
  analysis to trace back to their root cause.
- Ported FilterText.dll to x64 (Goddess loads it at startup for
  name filtering).

Verified by actually launching Goddess, Bishop, S3Relay and
GameServer together from the x64 build output and confirming they
stay up and fully interconnect, not just that they compile.

Includes the x64 build artifacts (SwordOnline/Lib/x64/, including a
user-built Berkeley DB 6.2.32 x64) following this repo's existing
convention of checking in prebuilt dependency libraries.

+++ SwordOnline/Headers/KProtocol.h
@@ -846,7 +846,7 @@ struct tagNewDelRoleResponse : public tagDBSelPlayer
 // 2003.05.11
 struct tagDBSyncPlayerInfo : public tagProtoHeader
 {
-	size_t	dataLength;
+	DWORD	dataLength;
 	char	szData[0];
 };
 
@@ -881,7 +881,7 @@ struct tagGuidableInfo : public tagProtoHeader
 	GUID guid;
 	WORD nExtPoint;			//可用的附送点
 	WORD nChangePoint;		//变化的附送点
-	size_t	datalength;
+	DWORD	datalength;
 	char	szData[0];
 
 };
@@ -1110,7 +1110,7 @@ struct tagRegisterFamily : public tagProtoHeader
  */
 struct tagGS2GWSaveRole : public tagProtoHeader
 {
-	size_t	datalength;
+	DWORD	datalength;
 	BYTE	szData[0];
 };
 
@@ -1140,7 +1140,7 @@ typedef struct
 struct TProcessData
 {
 	unsigned char	nProtoId;
-	size_t			nDataLen;//TRoleNetMsg时表示该Block的实际数据长度,TProcessData时表示Stream的实际数据长度
+	DWORD			nDataLen;//TRoleNetMsg时表示该Block的实际数据长度,TProcessData时表示Stream的实际数据长度
 	unsigned long	ulIdentity;
 	bool			bLeave;
 	char			szName[32];
